plugins/databricks-engineering/templates/data-product/requirements.txt
# Data Product Dependencies

pyspark==3.5.0
delta-spark==3.0.0
pyyaml==6.0.1
pytest==7.4.3
databricks-sdk==0.18.0
